# By Jorge (1) and jcvicente (1)
# Via Jorge (1) and jcvicente (1)
* 'master' of https://github.com/captainvera/CGProj:
  Fixed Table Texture Mapping
  Spotlight on car

# Conflicts:
#	src/Car.h
#	src/GameManager.cpp

+++ src/Vector3.h
@@ -12,7 +12,7 @@
 class Vector3 {
 private:
 public:
-	GLdouble _x,
+	GLfloat _x,
 			 _y,
 			 _z;
 	Vector3();
